decouple UI code and others
