+++ .gitignore
@@ -77,3 +77,5 @@ deps/cystdf/_cystdf.cpython-39-darwin.so
 deps/cystdf/_cystdf.cpython-39-darwin.so
 deps/cystdf/cystdf_amalgamation.c
 deps/cystdf/cython_test.py
+STDF-Viewer.config
+test.py
